# [MIR] SwitchInt Everywhere