@nikomatsakis 2017-02-02
So my take on this has traditionally been that I would prefer to do it as a kind of LOWERING step. That is, we would retain the semantic information when we build the tree, but lose it after dropck. However, I am not sure that this has a lot of value, and I'd be willing to go forward with this PR (after I give it a bit more of a look...), with the understanding that if, at some time later, we want to reason about the fact that some particular integer represents an enum discriminant, we *may* modify MIR to make that more direct (of course, one could also imagine tracking that fact via some kind of limited abstract interpretation or other analysis).