@arielb1 2017-02-02
Could you add a `TerminatorKind::if` constructor?