@nagisa 2017-02-04
I feel like this is pretty much done ± the fixes for test failures. If the travis passes, its ready to be r+d.