@nikomatsakis 2017-02-07
So, I'm happy with this PR, but I want to give @rust-lang/compiler a chance to weigh in. I'll r+ a litle later today, unless anyone raises any blocking objections in the next few hours. =)