@nikomatsakis 2017-02-07
@bors r+ 

But @eddyb and @pnkfelix have a side-condition: @eddyb stipulates that you must change the name to `Switch`. And @pnkfelix that you can't do it now so as not to disturb "in-flight" PRs. So you have to come back and do it later. :)