@nagisa 2017-02-07
> And @pnkfelix that you can't do it now so as not to disturb "in-flight" PRs. So you have to come back and do it later. :)

A commit to rename is already a part of this PR (last commit). Should I remove the commit before this PR lands?