@nikomatsakis 2017-02-08
@nagisa -- one thing. On discussing with @stoklund, we were saying that LLVM's generated IR will be more optimal if we have a *precise* match. So we should at minimum open a FIXME that we make `SwitchInt` only optionally have an "otherwise" case.