@nagisa 2017-02-08
@nikomatsakis counter-comment: technically MIR SwitchInt can and could for a long time represent pretty much anything that LLVM can. You cannot leave out the target for default branch in LLVM’s `switch` instruction either. So there’s two alternatives for generated LLVM IR (which one of them is the so-called *precise* one?):

```
switch i32 %_0, label %bbWithUnreachable [ i32 X, label %bbX
                                                                        i32 Y, label %bbY ]
```

or an equivalent

```
switch i32 %_0, label %bbY [ i32 X, label %bbX ]
```

Now I’m not sure which one is more optimal from the LLVM standpoint, but you can easily go from the 1st IR to the 2nd IR by popping 1 value from the list of values which are being switched on, and also back from 2nd to the 1st by pushing the corresponding value.

I’d myself argue that precise SwitchInt is much more interesting before translation to LLVM-IR, honestly as it may open some interesting semantic analysis (though these could be done before MIR as well).