@frewsxcv 2017-02-09
Just did a rollup [here](https://github.com/rust-lang/rust/pull/39663). Your PR here is the only substantive one included in the rollup. The rollup failed twice (after a retry) with the same error:

https://ci.appveyor.com/project/rust-lang/rust/build/1.0.1869/job/k00gvxe07n4mfql7

Thoughts? I'm going to take this off the queue just so it doesn't run again.

@bors r-