@stoklund 2017-02-09
@nagisa It's usually better to give LLVM as much information as possible, so use the switch with an unreachable default case:

```llvm
switch i32 %_0, label %bbWithUnreachable [ i32 X, label %bbX
                                           i32 Y, label %bbY ]
```

Concretely, this means that if the switch gets lowered to jump table, LLVM can skip the bounds check before loading from the jump table.