@nagisa 2017-02-10
Should be okay for 'nother go at bots.