@eddyb 2017-02-12
@bors r=nikomatsakis