@alexcrichton 2017-02-13
@bors: r=nikomatsakis